# docs/openapi-monolith-seed.yaml
openapi: 3.0.3
info:
  title: Ledgerfit
  version: 0.1.0
  description: >-
    Permissioned-ledger ML control plane for multi-pharmacy controlled-substance
    surveillance: train association and risk models from immutable prescription
    events with smart-contract-triggered retrain and score cycles.
  x-product-slug: ai-decentralized-ledgerfit
  x-source: ai-in-decentralized+ai/2019-03-blockchain-ml-taowang-v2-190322201510/
servers:
  - url: https://api.example.com
...
paths:
...
  /v1/members/{memberId}/freeze:
    parameters:
      - $ref: "#/components/parameters/MemberId"
    post:
      operationId: freezePharmacyMember
      tags: [LedgersDataSources]
      summary: Freeze outbound anchoring and scoring for a pharmacy member
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreezeOrderCreate"
      responses:
        "201":
          description: Freeze issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeFreezeOrder"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    MemberId:
      name: memberId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    FreezeOrder:
      type: object
      required: [id, memberId, reason, issuedBy, issuedAt]
      properties:
        id:
          type: string
        memberId:
          type: string
        reason:
          type: string
        issuedBy:
          type: string
        issuedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
    FreezeOrderCreate:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
        expiresAt:
          type: string
          format: date-time
...
    DataEnvelopeFreezeOrder:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/FreezeOrder"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...