# platform/api-server/src/domains/ledgers/handlers/freeze.handlers.ts
/**
 * Handlers for Freeze
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: ledgers
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { LedgersDomainModule } from "../dependencies/ledgers-ddd.dependencies.js";
import type { FreezePharmacyMemberInput } from "@ledgerfit/services/ledgers";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for POST /v1/members/{memberId}/freeze
 * Operation: freezePharmacyMember
 */
export async function freezePharmacyMember(
  request: FastifyRequest<{ Params: { memberId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: LedgersDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: FreezePharmacyMemberInput = {
    memberId: request.params.memberId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.freezes.get.execute(input);
  return reply.code(201).send(result);
}